add nonlinux interface implementations
