+++ nmproxy/router/firewall.go
@@ -3,22 +3,34 @@ package router
 import (
 	"context"
 
-	"github.com/coreos/go-iptables/iptables"
 	"github.com/gravitl/netmaker/logger"
 	"github.com/gravitl/netmaker/models"
 )
 
 var (
-	fwCrtl  firewallController
-	running bool
+	fwCrtl firewallController
 )
+
+type rulesCfg struct {
+	isIpv4   bool
+	rulesMap map[string][]ruleInfo
+}
+
+type ruleInfo struct {
+	rule  []string
+	table string
+	chain string
+}
+type ruletable map[string]rulesCfg
+
+type serverrulestable map[string]ruletable
 
 const (
 	ingressTable = "ingress"
 )
 
 type firewallController interface {
-	// CreateChains  creates a firewall chains and default rules
+	// CreateChains  creates a firewall chains and jump rules
 	CreateChains() error
 	// InsertIngressRoutingRules inserts a routing firewall rules for ingressGW
 	InsertIngressRoutingRules(server string, r models.ExtClientInfo) error
@@ -28,7 +40,7 @@ type firewallController interface {
 	RemoveRoutingRules(server, tableName, peerKey string) error
 	// DeleteRoutingRule removes rules related to a peer
 	DeleteRoutingRule(server, tableName, srcPeer, dstPeer string) error
-	// CleanRoutingRules cleans a firewall set of containers
+	// CleanRoutingRules cleans a firewall set of containers related to a server
 	CleanRoutingRules(server, tableName string)
 	// FetchRules - fetches current state of rules from controller
 	FetchRuleTable(server string, ruleTableName string) ruletable
@@ -37,14 +49,6 @@ type firewallController interface {
 	// FlushAll - clears all rules from netmaker chains and deletes the chains
 	FlushAll()
 }
-
-type rulesCfg struct {
-	isIpv4   bool
-	rulesMap map[string][]ruleInfo
-}
-type ruletable map[string]rulesCfg
-
-type serverrulestable map[string]ruletable
 
 func Init(ctx context.Context) error {
 	logger.Log(0, "Starting firewall...")
@@ -55,30 +59,4 @@ func Init(ctx context.Context) error {
 		fwCrtl.FlushAll()
 	}()
 	return fwCrtl.CreateChains()
-}
-
-// newFirewall if supported, returns an iptables manager, otherwise returns a nftables manager
-func newFirewall(parentCTX context.Context) firewallController {
-
-	var manager firewallController
-	if isIptablesSupported() {
-		logger.Log(0, "iptables is supported")
-		ipv4Client, _ := iptables.NewWithProtocol(iptables.ProtocolIPv4)
-		ipv6Client, _ := iptables.NewWithProtocol(iptables.ProtocolIPv6)
-		ctx, cancel := context.WithCancel(parentCTX)
-		manager = &iptablesManager{
-			ctx:        ctx,
-			stop:       cancel,
-			ipv4Client: ipv4Client,
-			ipv6Client: ipv6Client,
-			ingRules:   make(serverrulestable),
-		}
-		return manager
-	}
-
-	logger.Log(0, "iptables is not supported, using nftables")
-
-	// TODO - nft table manager
-
-	return manager
 }
